--- a/.github/workflows/actions.yml
+++ b/.github/workflows/actions.yml
@@ -2,7 +2,7 @@ name: Tests
 
 on:
   push:
-    branches: [ jax-tracing-fix ]
+    branches: [ master ]
   pull_request:
   release:
     types: [created]