--- a/keras/src/backend/jax/trainer.py
+++ b/keras/src/backend/jax/trainer.py
@@ -1,6 +1,5 @@
 import collections
 import itertools
-from functools import partial
 
 import jax
 import numpy as np
@@ -224,108 +223,109 @@ class JAXTrainer(base_trainer.Trainer):
         )
         return outputs, non_trainable_variables
 
+    def _make_function(self, step_function, concatenate_outputs=False):
+        if self.steps_per_execution > 1:
+            if concatenate_outputs:
+
+                def concatenate(outputs):
+                    output = outputs[0]
+                    for next_output in outputs[1:]:
+                        output = tree.map_structure(
+                            lambda t1, t2: jax.numpy.concatenate([t1, t2]),
+                            output,
+                            next_output,
+                        )
+                    return output
+
+                if not self.run_eagerly and self.jit_compile:
+                    concatenate = jax.jit(concatenate)
+
+                def function(state, iterator):
+                    data = next(iterator)
+                    outputs, state = step_function(state, data)
+                    outputs = [outputs]
+                    try:
+                        for _ in range(self.steps_per_execution - 1):
+                            data = next(iterator)
+                            _outputs, state = step_function(state, data)
+                            outputs.append(_outputs)
+                    except StopIteration:
+                        pass
+                    outputs = concatenate(outputs)
+                    return outputs, state
+
+            else:
+
+                def function(state, iterator):
+                    data = next(iterator)
+                    outputs, state = step_function(state, data)
+                    try:
+                        for _ in range(self.steps_per_execution - 1):
+                            data = next(iterator)
+                            outputs, state = step_function(state, data)
+                    except StopIteration:
+                        pass
+                    return outputs, state
+
+        else:
+
+            def function(state, iterator):
+                return step_function(state, next(iterator))
+
+        return function
+
     def make_train_function(self, force=False):
         if self.train_function is not None and not force:
             return
-
-        def one_train_step(state, data):
-            data = data[0]
-            return self.train_step(state, data)
-
-        def multi_train_steps(state, data):
-            for single_step_data in data:
-                logs, state = one_train_step(state, [single_step_data])
-            return logs, state
-
-        if self.steps_per_execution > 1:
-            train_step = multi_train_steps
-        else:
-            train_step = one_train_step
-
         if not self.run_eagerly and self.jit_compile:
-            # Note that we mark the state and data to be donated to jax,
+            # Note that we mark the state to be donated to jax,
             # so that jax will reuse the memory buffer for outputs.
             # This will reduce the memory usage of the training function by
             # half.
-            @partial(jax.jit, donate_argnames="state")
-            def compiled_train_step(state, data):
-                return train_step(state, data)
-
-            self.train_function = compiled_train_step
-
+            train_step = jax.jit(self.train_step, donate_argnums=0)
         else:
-            self.train_function = train_step
+            train_step = self.train_step
+
+        step_function = self._make_function(train_step)
+
+        self.train_function = step_function
 
     def make_test_function(self, force=False):
         if self.test_function is not None and not force:
             return
-
-        def one_test_step(state, data):
-            data = data[0]
-            return self.test_step(state, data)
-
-        def multi_test_steps(state, data):
-            for single_step_data in data:
-                logs, state = one_test_step(state, [single_step_data])
-            return logs, state
-
-        if self.steps_per_execution > 1:
-            test_step = multi_test_steps
-        else:
-            test_step = one_test_step
-
         if not self.run_eagerly and self.jit_compile:
-            # Note that we mark the state and data to be donated to jax,
+            # Note that we mark the state to be donated to jax,
             # so that jax will reuse the memory buffer for outputs.
             # This will reduce the memory usage of the training function by
             # half.
-            @partial(jax.jit, donate_argnames="state")
-            def compiled_test_step(state, data):
-                return test_step(state, data)
-
-            self.test_function = compiled_test_step
-
+            test_step = jax.jit(self.test_step, donate_argnums=0)
         else:
-            self.test_function = test_step
+            test_step = self.test_step
+
+        step_function = self._make_function(test_step)
+
+        self.test_function = step_function
 
     def make_predict_function(self, force=False):
         if self.predict_function is not None and not force:
             return self.predict_function
 
-        def one_predict_step(state, data):
-            data = data[0]
-            return self.predict_step(state, data)
-
-        def multi_predict_steps(state, data):
-            outputs, trainable_variables = one_predict_step(state, data[:1])
-
-            for single_step_data in data[1:]:
-                step_outputs, trainable_variables = one_predict_step(
-                    state,
-                    [single_step_data],
-                )
-                outputs = tree.map_structure(
-                    lambda t1, t2: jax.numpy.concatenate([t1, t2]),
-                    outputs,
-                    step_outputs,
-                )
-            return outputs, trainable_variables
-
-        if self.steps_per_execution > 1:
-            predict_step = multi_predict_steps
-        else:
-            predict_step = one_predict_step
+        def predict_step(state, data):
+            outputs, non_trainable_variables = self.predict_step(state, data)
+            return outputs, (state[0], non_trainable_variables)
 
         if not self.run_eagerly and self.jit_compile:
+            predict_step = jax.jit(predict_step)
 
-            @jax.jit
-            def compiled_predict_step(state, data):
-                return predict_step(state, data)
+        _step_function = self._make_function(
+            predict_step, concatenate_outputs=True
+        )
 
-            self.predict_function = compiled_predict_step
+        def step_function(state, iterator):
+            outputs, state = _step_function(state, iterator)
+            return outputs, state[1]
 
-        else:
-            self.predict_function = predict_step
+        self.predict_function = step_function
 
     @traceback_utils.filter_traceback
     def fit(
@@ -406,45 +406,46 @@ class JAXTrainer(base_trainer.Trainer):
             callbacks.on_epoch_begin(epoch)
 
             self._jax_state_synced = True
-            for step, data in epoch_iterator:
-                # Callbacks
-                callbacks.on_train_batch_begin(step)
+            with epoch_iterator.catch_stop_iteration():
+                for step, iterator in epoch_iterator:
+                    # Callbacks
+                    callbacks.on_train_batch_begin(step)
 
-                # Train step
-                if self._jax_state_synced:
-                    # The state may have been synced by a callback.
-                    state = self._get_jax_state(
-                        trainable_variables=True,
-                        non_trainable_variables=True,
-                        optimizer_variables=True,
-                        metrics_variables=True,
-                        purge_model_variables=True,
-                    )
-                    self._jax_state_synced = False
+                    # Train step
+                    if self._jax_state_synced:
+                        # The state may have been synced by a callback.
+                        state = self._get_jax_state(
+                            trainable_variables=True,
+                            non_trainable_variables=True,
+                            optimizer_variables=True,
+                            metrics_variables=True,
+                            purge_model_variables=True,
+                        )
+                        self._jax_state_synced = False
 
-                logs, state = self.train_function(state, data)
-                (
-                    trainable_variables,
-                    non_trainable_variables,
-                    optimizer_variables,
-                    metrics_variables,
-                ) = state
+                    logs, state = self.train_function(state, iterator)
+                    (
+                        trainable_variables,
+                        non_trainable_variables,
+                        optimizer_variables,
+                        metrics_variables,
+                    ) = state
 
-                # Setting _jax_state enables callbacks to force a state sync
-                # if they need to.
-                self._jax_state = {
-                    "trainable_variables": trainable_variables,
-                    "non_trainable_variables": non_trainable_variables,
-                    "optimizer_variables": optimizer_variables,
-                    "metrics_variables": metrics_variables,
-                }
-                # Dispatch callbacks. This takes care of async dispatch.
-                callbacks.on_train_batch_end(step, logs)
+                    # Setting _jax_state enables callbacks to force a state sync
+                    # if they need to.
+                    self._jax_state = {
+                        "trainable_variables": trainable_variables,
+                        "non_trainable_variables": non_trainable_variables,
+                        "optimizer_variables": optimizer_variables,
+                        "metrics_variables": metrics_variables,
+                    }
+                    # Dispatch callbacks. This takes care of async dispatch.
+                    callbacks.on_train_batch_end(step, logs)
 
-                if self.stop_training:
-                    # Stop training if a callback has set
-                    # this flag in on_(train_)batch_end.
-                    break
+                    if self.stop_training:
+                        # Stop training if a callback has set
+                        # this flag in on_(train_)batch_end.
+                        break
 
             # Reattach state to the model (if not already done by a callback).
             # NOTE: doing this after each step would be a big performance
@@ -562,41 +563,42 @@ class JAXTrainer(base_trainer.Trainer):
         self.reset_metrics()
 
         self._jax_state_synced = True
-        for step, data in epoch_iterator:
-            callbacks.on_test_batch_begin(step)
+        with epoch_iterator.catch_stop_iteration():
+            for step, iterator in epoch_iterator:
+                callbacks.on_test_batch_begin(step)
 
-            if self._jax_state_synced:
-                # The state may have been synced by a callback.
-                state = self._get_jax_state(
-                    trainable_variables=True,
-                    non_trainable_variables=True,
-                    metrics_variables=True,
-                    purge_model_variables=True,
-                )
-                self._jax_state_synced = False
+                if self._jax_state_synced:
+                    # The state may have been synced by a callback.
+                    state = self._get_jax_state(
+                        trainable_variables=True,
+                        non_trainable_variables=True,
+                        metrics_variables=True,
+                        purge_model_variables=True,
+                    )
+                    self._jax_state_synced = False
 
-            logs, state = self.test_function(state, data)
-            (
-                trainable_variables,
-                non_trainable_variables,
-                metrics_variables,
-            ) = state
+                logs, state = self.test_function(state, iterator)
+                (
+                    trainable_variables,
+                    non_trainable_variables,
+                    metrics_variables,
+                ) = state
 
-            # Setting _jax_state enables callbacks to force a state sync
-            # if they need to.
-            self._jax_state = {
-                # I wouldn't recommend modifying non-trainable model state
-                # during evaluate(), but it's allowed.
-                "trainable_variables": trainable_variables,
-                "non_trainable_variables": non_trainable_variables,
-                "metrics_variables": metrics_variables,
-            }
+                # Setting _jax_state enables callbacks to force a state sync
+                # if they need to.
+                self._jax_state = {
+                    # I wouldn't recommend modifying non-trainable model state
+                    # during evaluate(), but it's allowed.
+                    "trainable_variables": trainable_variables,
+                    "non_trainable_variables": non_trainable_variables,
+                    "metrics_variables": metrics_variables,
+                }
 
-            # Dispatch callbacks. This takes care of async dispatch.
-            callbacks.on_test_batch_end(step, logs)
+                # Dispatch callbacks. This takes care of async dispatch.
+                callbacks.on_test_batch_end(step, logs)
 
-            if self.stop_evaluating:
-                break
+                if self.stop_evaluating:
+                    break
 
         # Reattach state back to model (if not already done by a callback).
         self.jax_state_sync()
@@ -627,13 +629,15 @@ class JAXTrainer(base_trainer.Trainer):
 
         if not all(layer.built for layer in self._flatten_layers()):
             # Build the model on one batch of data.
-            for _, data in epoch_iterator:
+            for _, iterator in epoch_iterator:
                 # Build model
-                x, _, _ = data_adapter_utils.unpack_x_y_sample_weight(data[0])
+                x, _, _ = data_adapter_utils.unpack_x_y_sample_weight(
+                    next(iterator)
+                )
                 with backend.StatelessScope():
                     self(x)
                 break
-
+            epoch_iterator.reset()
         # Container that configures and calls callbacks.
         if not isinstance(callbacks, callbacks_module.CallbackList):
             callbacks = callbacks_module.CallbackList(
@@ -669,28 +673,29 @@ class JAXTrainer(base_trainer.Trainer):
         self._jax_state_synced = True
         outputs = None
         non_trainable_variables = None
-        for step, x in epoch_iterator:
-            callbacks.on_predict_batch_begin(step)
-            if self._jax_state_synced:
-                # The state may have been synced by a callback.
-                state = self._get_jax_state(
-                    trainable_variables=True,
-                    non_trainable_variables=True,
+        with epoch_iterator.catch_stop_iteration():
+            for step, iterator in epoch_iterator:
+                callbacks.on_predict_batch_begin(step)
+                if self._jax_state_synced:
+                    # The state may have been synced by a callback.
+                    state = self._get_jax_state(
+                        trainable_variables=True,
+                        non_trainable_variables=True,
+                    )
+                    self._purge_model_variables(non_trainable_variables=True)
+                    self._jax_state_synced = False
+                else:
+                    state = (state[0], non_trainable_variables)
+                batch_outputs, non_trainable_variables = self.predict_function(
+                    state, iterator
                 )
-                self._purge_model_variables(non_trainable_variables=True)
-                self._jax_state_synced = False
-            else:
-                state = (state[0], non_trainable_variables)
-            batch_outputs, non_trainable_variables = self.predict_function(
-                state, x
-            )
-            outputs = append_to_outputs(batch_outputs, outputs)
+                outputs = append_to_outputs(batch_outputs, outputs)
 
-            # Dispatch callbacks. This takes care of async dispatch.
-            callbacks.on_predict_batch_end(step, {"outputs": batch_outputs})
+                # Dispatch callbacks. This takes care of async dispatch.
+                callbacks.on_predict_batch_end(step, {"outputs": batch_outputs})
 
-            if self.stop_predicting:
-                break
+                if self.stop_predicting:
+                    break
 
         self._jax_state = {
             # I wouldn't recommend modifying non-trainable model state
@@ -722,11 +727,12 @@ class JAXTrainer(base_trainer.Trainer):
             sample_weight = data_adapter_utils.class_weight_to_sample_weights(
                 y, class_weight
             )
-        data = (x, y, sample_weight)
-        data = _distribute_data(data)
+
+        def data():
+            yield _distribute_data((x, y, sample_weight))
 
         # Maybe build model
-        self._symbolic_build(data_batch=data)
+        self._symbolic_build(data_batch=next(data()))
         self._record_training_state_sharding_spec()
         self.make_train_function()
 
@@ -739,7 +745,7 @@ class JAXTrainer(base_trainer.Trainer):
             purge_model_variables=False,
         )
         self._jax_state_synced = False
-        logs, state = self.train_function(state, [data])
+        logs, state = self.train_function(state, data())
 
         # State sync
         (
@@ -771,10 +777,11 @@ class JAXTrainer(base_trainer.Trainer):
     ):
         self._assert_compile_called("test_on_batch")
 
-        data = (x, y, sample_weight)
-        data = _distribute_data(data)
+        def data():
+            yield _distribute_data((x, y, sample_weight))
+
         # Maybe build model
-        self._symbolic_build(data_batch=data)
+        self._symbolic_build(data_batch=next(data()))
         self._record_training_state_sharding_spec()
         self.make_test_function()
 
@@ -786,7 +793,7 @@ class JAXTrainer(base_trainer.Trainer):
             purge_model_variables=False,
         )
         self._jax_state_synced = False
-        logs, state = self.test_function(state, [data])
+        logs, state = self.test_function(state, data())
 
         # State sync
         trainable_variables, non_trainable_variables, metrics_variables = state
@@ -818,8 +825,12 @@ class JAXTrainer(base_trainer.Trainer):
             purge_model_variables=False,
         )
         self._jax_state_synced = False
+
+        def data():
+            yield (x,)
+
         batch_outputs, non_trainable_variables = self.predict_function(
-            state, [(x,)]
+            state, data()
         )
         self._jax_state = {
             "non_trainable_variables": non_trainable_variables,
@@ -992,6 +1003,9 @@ def _distribute_data(data, layouts=None):
 
 
 class JAXEpochIterator(EpochIterator):
+    def __next__(self):
+        return next(self._epoch_iterator)
+
     def _get_iterator(self):
         distribution = distribution_lib.distribution()
         if distribution is not None: