Fixes https://github.com/keras-team/keras/issues/20402
Fixes https://github.com/keras-team/keras/issues/20411

The solution uses an approach similar to TensorFlow with `EpochIterator` returning an iterator instead of a list of batches. This prevents data loading bottlenecks when `steps_per_execution` > 1. It is enough to have a non-jitted for loop that iterates over the execution steps and makes calls to the inner `train/test/predict_step` function thanks to Jax's asynchronous dispatch.

This [benchmark](https://gist.github.com/nicolaspi/6c6483abffb57dec3db79a6e4c1efa1f) shows constant compilation time and memory usage with an improvement for prediction speed:

Before the changes:
```
steps_per_execution = 1, model retracing count: 1, memory usage: 887MB, compilation overhead 1.25 seconds, inference speed: 0.284 ms/sample.
steps_per_execution = 5, model retracing count: 5, memory usage: 1638MB, compilation overhead 3.88 seconds, inference speed: 0.310 ms/sample.
steps_per_execution = 10, model retracing count: 10, memory usage: 3211MB, compilation overhead 7.32 seconds, inference speed: 0.377 ms/sample.
```

After the changes:
```
steps_per_execution = 1, model retracing count: 1, memory usage: 887MB, compilation overhead 1.38 seconds, inference speed: 0.281 ms/sample.
steps_per_execution = 5, model retracing count: 1, memory usage: 887MB, compilation overhead 1.96 seconds, inference speed: 0.277 ms/sample.
steps_per_execution = 10, model retracing count: 1, memory usage: 887MB, compilation overhead 2.07 seconds, inference speed: 0.273 ms/sample.
steps_per_execution = 50, model retracing count: 1, memory usage: 887MB, compilation overhead 2.01 seconds, inference speed: 0.276 ms/sample.
steps_per_execution = 100, model retracing count: 1, memory usage: 887MB, compilation overhead 0.81 seconds, inference speed: 0.276 ms/sample.
```

This [benchmark](https://colab.research.google.com/gist/nicolaspi/ef2c37efb7f9c9bbcdf2fcfbb6d45ae6/untitled5.ipynb#scrollTo=gzRyT4VrhOJW) for training also shows an improvement in speed:

Before the changes (`steps_per_execution=1`; higher values gave slower training):
```
Epoch 1/40
160/160 ━━━━━━━━━━━━━━━━━━━━ 49s 143ms/step - accuracy: 0.0101 - loss: 4.9874 - val_accuracy: 0.0081 - val_loss: 4.7888
Epoch 2/40
160/160 ━━━━━━━━━━━━━━━━━━━━ 20s 86ms/step - accuracy: 0.0260 - loss: 4.6006 - val_accuracy: 0.0083 - val_loss: 4.8747
Epoch 3/40
160/160 ━━━━━━━━━━━━━━━━━━━━ 14s 86ms/step - accuracy: 0.0392 - loss: 4.3919 - val_accuracy: 0.0083 - val_loss: 5.1281
```

After the changes (`steps_per_execution=32`):
```
Epoch 1/40
160/160 ━━━━━━━━━━━━━━━━━━━━ 49s 90ms/step - accuracy: 0.0122 - loss: 4.9216 - val_accuracy: 0.0095 - val_loss: 4.7878
Epoch 2/40
160/160 ━━━━━━━━━━━━━━━━━━━━ 20s 72ms/step - accuracy: 0.0298 - loss: 4.5174 - val_accuracy: 0.0094 - val_loss: 4.9087
Epoch 3/40
160/160 ━━━━━━━━━━━━━━━━━━━━ 14s 72ms/step - accuracy: 0.0423 - loss: 4.3515 - val_accuracy: 0.0094 - val_loss: 5.2104
```